ProjMPSs

Documentation for ProjMPSs.

ProjMPSs.BlockedMPS

ProjMPSs.ProjMPS

ProjMPSs.Projector

Base.:&

Base.:+

Base.:<

Base.:|

Base.getindex

Base.keys

Base.length

Base.values

ITensors.SiteTypes.siteinds

LinearAlgebra.norm

ProjMPSs.extractdiagonal

ProjMPSs.projcontract

ProjMPSs.projcontract

Quantics.automul

Quantics.makesitediagonal

Quantics.rearrange_siteinds

ProjMPSs.BlockedMPS — Type

BlockedMPS is a structure that holds multiple MPSs (ProjMPS) that are associated with different non-overlapping projectors.

ProjMPSs.ProjMPS — Type

An MPS with a projector.

ProjMPSs.Projector — Type

A projector represents a projection of a tensor from a set of its indices to integers. Each index is projected to a positive integer.

Base.:& — Method

a & b represents the intersection of the indices that a and b are projected at.

Base.:+ — Method

Add two BlockedMPS objects.

If the two projects have the same projectors in the same order, the resulting BlockedMPS will have the same projectors in the same order. By default, we use directsum algorithm to compute the sum and no truncation is performed.

Base.:< — Method

a < b means that a is projected at a subset of the indices that b is projected at.

Base.:| — Method

a | b represents the union of the indices that a and b are projected at.

If a is projected at inds=1 and b is not projected for the same inds, then a | b is not projected for inds.

Base.getindex — Method

Indexing for BlockedMPS. This is deprecated and will be removed in the future.

Base.keys — Method

Return the keys, i.e., projectors of the BlockedMPS.

Base.length — Method

Get the number of sites in the BlockedMPS

Base.values — Method

Return the values, i.e., ProjMPS of the BlockedMPS.

ITensors.SiteTypes.siteinds — Method

Return the site indices of the BlockedMPS. The site indices are returned as a vector of sets, where each set corresponds to the site indices at each site.

LinearAlgebra.norm — Method

Return the norm of the BlockedMPS.

ProjMPSs.extractdiagonal — Method

Extract diagonal of the BlockedMPS for s, s', ... for a given site index s, where s must have a prime level of 0.

ProjMPSs.projcontract — Method

Project two ProjMPS objects to proj before contracting them. The results are summed.

ProjMPSs.projcontract — Method

Project two ProjMPS objects to proj before contracting them.

Quantics.automul — Method

By default, elementwise multiplication will be performed.

Quantics.makesitediagonal — Method

Make the BlockedMPS diagonal for a given site index s by introducing a dummy index s'.

Quantics.rearrange_siteinds — Method

Rearrange the site indices of the BlockedMPS according to the given order. If nessecary, tensors are fused or split to match the new order.