ProjMPSs

Documentation for ProjMPSs.

ProjMPSs.BlockedMPS

ProjMPSs.LazyContraction

ProjMPSs.ProjMPS

ProjMPSs.Projector

ProjMPSs.Projector

Base.:&

Base.:+

Base.:<

Base.:|

Base.getindex

Base.isdisjoint

Base.keys

Base.length

Base.values

LinearAlgebra.norm

ProjMPSs._add_patching

ProjMPSs._next_projindex

ProjMPSs.adaptive_patching

ProjMPSs.adaptive_patching

ProjMPSs.adaptivecontract

ProjMPSs.add_patching

ProjMPSs.contract

ProjMPSs.extractdiagonal

ProjMPSs.makesitediagonal

ProjMPSs.projcontract

ProjMPSs.projcontract

ProjMPSs.project

ProjMPSs.rearrange_siteinds

ProjMPSs.siteindices

ProjMPSs.BlockedMPS — Type

BlockedMPS is a structure that holds multiple MPSs (ProjMPS) that are associated with different non-overlapping projectors.

ProjMPSs.LazyContraction — Type

Lazy evaluation for contraction of two ProjMPS objects.

ProjMPSs.ProjMPS — Type

An MPS with a projector.

ProjMPSs.Projector — Type

A projector represents a projection of a tensor from a set of its indices to integers. Each index is projected to a positive integer.

ProjMPSs.Projector — Method

Constructing a projector from a single pair of index and integer.

Base.:& — Method

a & b represents the intersection of the indices that a and b are projected at.

Base.:+ — Method

Add two BlockedMPS objects.

If the two projects have the same projectors in the same order, the resulting BlockedMPS will have the same projectors in the same order. By default, we use directsum algorithm to compute the sum and no truncation is performed.

Base.:< — Method

a < b means that a is projected at a subset of the indices that b is projected at.

Base.:| — Method

a | b represents the union of the indices that a and b are projected at.

If a is projected at inds=1 and b is not projected for the same inds, then a | b is not projected for inds.

Base.getindex — Method

Indexing for BlockedMPS. This is deprecated and will be removed in the future.

Base.isdisjoint — Method

Return if projectors are not overlapping

Base.keys — Method

Return the keys, i.e., projectors of the BlockedMPS.

Base.length — Method

Get the number of sites in the BlockedMPS

Base.values — Method

Return the values, i.e., ProjMPS of the BlockedMPS.

LinearAlgebra.norm — Method

Return the norm of the BlockedMPS.

ProjMPSs._add_patching — Method

Add multiple ProjMPS objects on the same projector.

If the bond dimension of the result reaches maxdim, perform patching recursively to reduce the bond dimension.

ProjMPSs._next_projindex — Method

Return the next index to be projected.

ProjMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a ProjMPS exceeds maxdim, perform patching.

ProjMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a ProjMPS exceeds maxdim, perform patching.

ProjMPSs.adaptivecontract — Function

Perform contruction of two BlockedMPS objects.

The ProjMPS objects of each BlockedMPS do not overlap with each other. This makes the algorithm much simpler

ProjMPSs.add_patching — Method

Add multiple BlockedMPS objects.

ProjMPSs.contract — Method

Contract two Blocked MPS objects.

At each site, the objects must share at least one site index.

ProjMPSs.extractdiagonal — Method

Extract diagonal of the BlockedMPS for s, s', ... for a given site index s, where s must have a prime level of 0.

ProjMPSs.makesitediagonal — Method

Make the BlockedMPS diagonal for a given site index s by introducing a dummy index s'.

ProjMPSs.projcontract — Method

Project two ProjMPS objects to proj before contracting them. The results are summed.

ProjMPSs.projcontract — Method

Project two ProjMPS objects to proj before contracting them.

ProjMPSs.project — Method

Project the LazyContraction object to prj before evaluating it.

This may result in projecting the external indices of a and b.

ProjMPSs.rearrange_siteinds — Method

Rearrange the site indices of the BlockedMPS according to the given order. If nessecary, tensors are fused or split to match the new order.

ProjMPSs.siteindices — Method

Return the site indices of the BlockedMPS. The site indices are returned as a vector of sets, where each set corresponds to the site indices at each site.